Python Interview Q & A
Are you an aspiring Python Developer? A career in Python has seen an upward trend in 2023, and you can be a part of the ever-so-growing community. So, if you are ready to indulge yourself in the pool of knowledge and be prepared for the upcoming python interview, then you are at the right place.

𝐏𝐲𝐭𝐡𝐨𝐧 𝐈𝐧𝐭𝐞𝐫𝐯𝐢𝐞𝐰 𝐐𝐮𝐞𝐬𝐭𝐢𝐨𝐧𝐬 𝐟𝐨𝐫 𝐅𝐫𝐞𝐬𝐡𝐞𝐫𝐬:

This section on Python Interview Questions for freshers covers 70+ questions that are commonly asked during the interview process. As a fresher, you may be new to the interview process; however, learning these questions will help you answer the interviewer confidently and ace your upcoming interview.

Tagging W3Schools.com, GeeksforGeeks, Great Learning, AlmaBetter & Edureka   for providing great Resources !

Also Tagging Pradeepchandra Reddy S C Garima Rungta Singhal  Mohini Shrivastav Vidhi Jain  ISarthak Khanna Madhuparna Bit  for a wider audience reach

✅𝐅𝐨𝐥𝐥𝐨𝐰 Siddhesh Kumbhar 📌
-𝐟𝐨𝐫 𝐦𝐨𝐫𝐞 𝐬𝐮𝐜𝐡 𝐚𝐦𝐚𝐳𝐢𝐧𝐠 𝐧𝐨𝐭𝐞𝐬 𝐚𝐧𝐝 𝐜𝐨𝐧𝐭𝐞𝐧𝐭𝐬.

#python #machinelearning #bigdata #deeplearning  
#datascientists #datascience #dataanalytics #datascientist  #dataanalyst #computerscience  #pythonprogramming
#decisionintelligence #programming #coding #programmer 
#datavisualization #analytics #data #digital #datavisualization #microsoft #powerbi #Tableau #artificialintelligence  #ai #technology #tech #git #github #Linkedin
#sql #sqlquery #mysql #sqlserver #sqldeveloper #sqlqueries #sqldatabase #dbms #databasemanagementsystem #database 
#fresher #hiring #tier3  #notes #interview #project
#share  #interviewpreparation #learning